requirements.txt
# Simulação de Eventos Discretos
simpy==4.1.1

# Processamento de Arquivos YAML
PyYAML==6.0.2
